refactor(ui): split App into view-routing hook and surface components
